fix: Correct audio capabilities across Embrava Blynclight variants
## Problem
The base EmbravaBase class incorrectly included audio methods (play_sound,
stop_sound, mute, unmute) that are not supported by the standard Blynclight
device. These audio capabilities are only available on Plus and Mini variants.

## Solution
- **Remove audio methods** from EmbravaBase (affects standard Blynclight)
- **Add audio methods** to BlynclightPlus class where they belong
- **Update BlynclightMini** to inherit from BlynclightPlus instead of EmbravaBase
- **Fix inheritance hierarchy** to properly reflect device capabilities

## Changes
- `EmbravaBase`: Removed play_sound, stop_sound, mute, unmute methods
- `BlynclightPlus`: Added audio capability methods
- `BlynclightMini`: Changed inheritance to extend BlynclightPlus

## Impact
- Standard Blynclight devices no longer expose non-functional audio methods
- Plus and Mini variants retain full audio functionality
- Inheritance hierarchy now accurately reflects hardware capabilities
- All tests pass, no breaking API changes for supported device features
