+++ src/busylight_core/vendors/embrava/embrava_base.py
@@ -62,35 +62,6 @@ class EmbravaBase(Light):
         with self.batch_update():
             self.state.dim = False
 
-    def play_sound(self, music: int = 0, volume: int = 1, repeat: bool = False) -> None:
-        """Play a sound on the device.
-
-        :param music: Music track number to play (0-7)
-        :param volume: Volume level (0-3)
-        :param repeat: Whether the music repeats
-        """
-        with self.batch_update():
-            self.state.repeat = repeat
-            self.state.play = True
-            self.state.music = music
-            self.state.mute = False
-            self.state.volume = volume
-
-    def stop_sound(self) -> None:
-        """Stop playing any currently playing sound."""
-        with self.batch_update():
-            self.state.play = False
-
-    def mute(self) -> None:
-        """Mute the device sound output."""
-        with self.batch_update():
-            self.state.mute = True
-
-    def unmute(self) -> None:
-        """Unmute the device sound output."""
-        with self.batch_update():
-            self.state.mute = False
-
     def flash(
         self,
         color: tuple[int, int, int],
